Dev wp1 ensemble generation

+++ main_inversion.py
@@ -19,7 +19,7 @@ import numpy as np
 import perturbation.inversion as inv
 from time import perf_counter
 from collections import OrderedDict
-
+import yaml
 import pandas as pd
 from datetime import date, timedelta, datetime
 import perturbation.utils as utils
@@ -36,12 +36,14 @@ if __name__=="__main__" :
     parser.add_argument('--ckpt_dir', type = str, 
                         default ='/scratch/mrmn/brochetc/GAN_2D/Exp_StyleGAN_final/Set_1/stylegan2_stylegan_dom_256_lat-dim_512_bs_4_0.002_0.002_ch-mul_2_vars_u_v_t2m_noise_True/Instance_14/models/000024.pt')
     # Real Data Directory - PATH to samples of the dataset
-    parser.add_argument('--real_data_dir', type = str, default='/scratch/mrmn/brochetc/GAN_2D/datasets_full_indexing/IS_1_1.0_0_0_0_0_0_256_large_lt_done/')
+    parser.add_argument('--real_data_dir', type = str, 
+                        default='/scratch/mrmn/brochetc/GAN_2D/datasets_full_indexing/IS_1_1.0_0_0_0_0_0_256_large_lt_done/')
     # Output Directory - PATH where the output of the inversion will be saved
     parser.add_argument('--output_dir',type = str, 
-                        default ='/scratch/mrmn/sanchezv/project/results/test/Ens_Perceptual_Loss_Random_VGG/Inversion_Perceptual_Loss_Random_VGG/')
+                        default ='/scratch/mrmn/sanchezv/project/results/_Ens_Perceptual_Random_VGG_Loss/Inversion_Perceptual_Random_VGG_Loss/')
     # Pack Directory - PATH where the packed ensembles will be saved
-    parser.add_argument("--pack_dir", type=str, default = '/scratch/mrmn/sanchezv/project/results/test/Ens_Perceptual_Loss_Random_VGG/Pack_Perceptual_Loss_Random_VGG/') # storing "packed" (normalized) real data
+    parser.add_argument("--pack_dir", type=str, 
+                        default = '/scratch/mrmn/sanchezv/project/results/_Ens_Perceptual_Random_VGG_Loss/Pack_Perceptual_Random_VGG_Loss/') # storing "packed" (normalized) real data
     
     # Dataset information
     parser.add_argument("--normalization", type=str, default="meanmax", choices=["minmax", "meanmax"])
@@ -49,7 +51,7 @@ if __name__=="__main__" :
     parser.add_argument('--mean_file', type=str, default='Mean_4_var.npy') # not used if minmax normalization
     parser.add_argument('--min_file', type=str, default='min_rr_log.npy')  # not used if meanmax normalization
     
-    parser.add_argument('--device', type=str, default='cuda:0' if torch.cuda.is_available() else 'cpu')
+    parser.add_argument('--device', type=str, default='cuda:0')
 
     ############################ INVERSION PARAMETERS #################    
 
@@ -69,25 +71,28 @@ if __name__=="__main__" :
     parser.add_argument("--noise_optimize", type=int, default=1, choices=[0,1], help="joint optimization of noise and latent code (1) or latent code optimization only (0)?")
 
     parser.add_argument('--pixel_loss_type', type=str, default='mse', choices = ['mse', 'mae'])
-
+    
     parser.add_argument("--lambda_noise", type=float, default=1.0, help="weight of the noise regularization")
-    parser.add_argument("--lambda_vgg", type=float, default=0.0, help="weight of the vgg (perceptual) loss")
+    parser.add_argument("--lambda_vgg", type=float, default=1.0, help="weight of the vgg (perceptual) loss")
     parser.add_argument("--lambda_pixel", type=float, default=1.0, help="weight of the (mae/mse) pixel loss")
 
+    parser.add_argument("--vgg_computation", type=str, default='sol1', choices = ['sol1', 'sol2', 'sol3', 'sol4', 'sol5'], 
+                        help="Either we compute layer by layer and member per member but we have to triple th einput to make it rgb or all in one (naive)")
+    parser.add_argument("--vgg_state_dict_path", type=str, default='/home/mrmn/sanchezv/project/code/styleganpnria/perturbation/vgg_weights/vgg16-397923af.pth', help="Insert a path")
     parser.add_argument("--vgg_style_layers", type=int, nargs='+', default=[], help="style layers to include in vgg loss computation")
-    parser.add_argument("--vgg_feature_layers", type=int, nargs='+', default=[], help="feature layers to include in vgg computation")
+    parser.add_argument("--vgg_feature_layers", type=int, nargs='+', default=[0,1,2,3], help="feature layers to include in vgg computation")
     parser.add_argument("--vgg_alpha_feature", type=float, default=1.0, help="weight of the feature/content loss")
     parser.add_argument("--vgg_alpha_style", type=float, default=0.01, help="weight of the style loss")
+    parser.add_argument("--vgg_loss_after_step", type=float, default=0, help="compute the vgg loss only after a given number of steps")
 
-    parser.add_argument("--invstep", type=int, default=1000, help="optimize iterations")
-    parser.add_argument("--inv_checkpoints", type=utils.str2intlist, default=[200,400,600,800,1000])
+    parser.add_argument("--invstep", type=int, default=2000, help="optimize iterations")
+    parser.add_argument("--inv_checkpoints", type=utils.str2intlist, default=[250,500,1000,1500,2000])
 
     ########################## CONTROL of Data to invert ######################
-
     parser.add_argument("--dates_file", type=str, default = 'Large_lt_test_labels.csv')
-    parser.add_argument("--date_start", type=str, default = "2021-06-01")
-    parser.add_argument("--date_stop", type=str, default = "2021-11-15")
-    parser.add_argument("--leadtimes", type=utils.str2intlist, default=[3,6,9,12,15,18,21,24,27,30,33,36,39,42,45])
+    parser.add_argument("--date_start", type=str, default = "2021-07-01")
+    parser.add_argument("--date_stop", type=str, default = "2021-08-31")
+    parser.add_argument("--leadtimes", type=utils.str2intlist, default= [3,6,9,12,15,18,21,24,27,30,33,36,39,42])
     
     parser.add_argument("--seed", type=int, default=42)
     
@@ -157,11 +162,11 @@ if __name__=="__main__" :
         latent_mean = torch.tensor(lm, dtype = torch.float32)
 
     ########### write inversion parameters to file ############
-    config_file = params.output_dir + "inversion_params.txt"
+    config_file = params.output_dir + "inversion_params.yaml"
     print("writing params config file:", config_file)
     try:
-        with open(config_file, 'w') as f_config:
-             f_config.write(json.dumps(params.__dict__))
+        file=open(config_file,"w")
+        yaml.dump(params.__dict__,file)
     except Exception as e:
          print("unable to write params config file")
          print(e)
@@ -171,38 +176,65 @@ if __name__=="__main__" :
     for key, value in params.__dict__.items():
         print(f"{key}: {value}")
 
+    if params.vgg_loss_after_step >= params.invstep:
+        print('The parameters vgg_loss_after_step cannot be superior or equal to the number of optim steps')
+        raise ValueError
 
     #################### main loop ##################
     for date_ in list_dates:
         print(date_)
         datename = date_.strftime('%Y-%m-%d')
         print("\n===========================")
-
         for lt in params.leadtimes:
-            print(datename,lt)
-            df0 = df_extract[(df_extract['Date']==date_) & (df_extract['LeadTime']==lt)]
-            if len(df0)==0:
-               print("# samples: 0")
-               continue
-
-            Ens_r = utils.load_batch_from_timestamp(df_extract, date_, lt, params.real_data_dir, Shape=params.Shape, var_indices=params.var_indices) #, crop_indices=params.crop_indices)
-
-            n_samples = np.min([Ens_r.shape[0], 6])
-            print(f"extracting {n_samples} samples for inversion\n")
-            Ens_r = Ens_r[:n_samples]
-
-            # normalise samples and save in pack dir. obs! make sure normalization is done correctly (according to how model was trained)
-            if params.normalization=="meanmax":
-               Ens_r = torch.tensor(0.95*(Ens_r - Means) / (Maxs), dtype = torch.float32)
-            elif params.normalization=="minmax":
-               Ens_r = torch.tensor(-1. + 2*(Ens_r - Mins) / (Maxs-Mins), dtype = torch.float32)
-            else:
-               raise ValueError(f"Unknown normalization: {params.normalization}")
-            np.save(params.pack_dir+f'Rsemble_{datename}_{lt}.npy', Ens_r.numpy().astype(np.float32))
-
             params.date_index = datename
             params.lt_index = lt
-            inv.optimize(Ens_r, G, latent_mean, params.device, params)
+            
+            # Check if the files already exists (to qave computation time)
+            already_exist = []
+            if os.path.isfile(params.pack_dir+f'Rsemble_{datename}_{lt}.npy'):
+                already_exist.append(True)
+            else :
+                already_exist.append(False)
+            for i in params.inv_checkpoints :
+                if os.path.isfile(params.output_dir+'w_{}_{}_{}.npy'.format(params.date_index,params.lt_index,i)):
+                    already_exist.append(True)
+                else :
+                    already_exist.append(False)
+                if os.path.isfile(params.output_dir+'invertFsemble_{}_{}_{}.npy'.format(params.date_index,params.lt_index,i)):
+                    already_exist.append(True)
+                else :
+                    already_exist.append(False)
+                if os.path.isfile(params.output_dir+'noise_{}_{}_{}.p'.format(params.date_index,params.lt_index,i)):
+                    already_exist.append(True)
+                else :
+                    already_exist.append(False)
+
+            if np.all(already_exist) :
+                print('The inversion was already done for the date {} with leadtime {}. This sample is skipped.'.format(datename,lt))
+            else :
+                print('Launching inversion process for the date {} with leadtime {}.'.format(datename,lt))
+                df0 = df_extract[(df_extract['Date']==date_) & (df_extract['LeadTime']==lt-1)]
+                if len(df0)==0:
+                   print("# samples: 0")
+                   continue
+
+                Ens_r = utils.load_batch_from_timestamp(df_extract, date_, lt, params.real_data_dir, Shape=params.Shape, var_indices=params.var_indices) #, crop_indices=params.crop_indices)
+
+                # n_samples = np.min([Ens_r.shape[0], 6])
+                # print(f"extracting {n_samples} samples for inversion\n")
+                # Ens_r = Ens_r[:n_samples]
+
+                # normalise samples and save in pack dir. obs! make sure normalization is done correctly (according to how model was trained)
+                if params.normalization=="meanmax":
+                   Ens_r = torch.tensor(0.95*(Ens_r - Means) / (Maxs), dtype = torch.float32)
+                elif params.normalization=="minmax":
+                   Ens_r = torch.tensor(-1. + 2*(Ens_r - Mins) / (Maxs-Mins), dtype = torch.float32)
+                else:
+                   raise ValueError(f"Unknown normalization: {params.normalization}")
+                np.save(params.pack_dir+f'Rsemble_{datename}_{lt}.npy', Ens_r.numpy().astype(np.float32))
+
+                
+                inv.optimize(Ens_r, G, latent_mean, params.device, params)
